restructure and setup for vscode/platformio
adding new onPacket method instead of weak references to callback functions. With this support it will be possible to receive a callback for any packet type on a selective basis or to receive all packets. Additional callback methods will be added to replace the other weak references still in place.
