+++ src/LocoNetUART.cpp
@@ -1,5 +1,5 @@
 #include "LocoNetUART.h"
-
+#ifndef ESP32
 void LocoNetUartClass::init()
 {
 }
@@ -12,3 +12,4 @@ LN_STATUS LocoNetUartClass::sendLocoNetPacketTry(lnMsg *txData, unsigned char uc
 	return LN_DONE;
 }
 
+#endif
