--- utility/component.mk
@@ -1,3 +1,0 @@
-#
-# Main Makefile. This is basically the same as a component makefile.
-#
